requirements.txt
aiohttp==3.9.3
aiosignal==1.2.0
anyio==4.2.0
appnope==0.1.2
apptools==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==2.0.4
comm==0.2.2
configobj==5.0.8
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
envisage==6.0.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.16.2
filelock==3.13.1
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2025.2.0
gmpy2==2.2.1
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipyevents==2.0.2
ipykernel==6.28.0
ipympl==0.9.3
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
kiwisolver==1.4.4
loguru==0.5.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mayavi==4.8.2.dev0
mistune==2.0.4
mlabtex==0.2.0
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.0.8
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.4
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.3.1
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyface==7.3.0
Pygments==2.17.2
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
scipy==1.12.0
Send2Trash==1.8.2
setuptools==68.2.2
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.13.3
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.2
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.2
traits==6.2.0
traitsui==7.2.1
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.1.0
vtk==9.2.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
widgetsnbextension==4.0.10
wslink==1.12.4
yarl==1.9.3
zipp==3.17.0
